# openapi.yaml
openapi: 3.0.0
info:
  title: Gym Training Registration API
  description: API for scheduling and managing gym training sessions
  version: 1.0.0
servers:
  - url: /
    description: API Server
paths:
  /trainees:
    get:
      security:
        - BearerAuth: [ ]
      tags:
        - Trainees
      summary: List all trainees
      description: Retrieve a paginated list of all trainees.
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
          description: The page number to retrieve.
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of trainees to return per page.
      responses:
        '200':
          description: List of trainees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trainee'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid pagination parameters"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Trainees
      summary: Create a new trainee
      description: Create a new trainee with a name, email, password, and optional timezone.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trainee'
            examples:
              trainee:
                summary: Example trainee registration
                value:
                  name: "John Doe"
                  email: "john.doe@example.com"
                  password: "securePassword123"
                  timezone: "Europe/Tallinn"
      responses:
        '201':
          description: Trainee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainee'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email is already in use"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trainees/{traineeId}:
    get:
      security:
        - BearerAuth: [ ]
      tags:
        - Trainees
      summary: Get trainee details
      description: Retrieve details of a specific trainee by their ID.
      parameters:
        - name: traineeId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the trainee to retrieve.
      responses:
        '200':
          description: Trainee details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainee'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      security:
        - BearerAuth: [ ]
      tags:
        - Trainees
      summary: Partially update a trainee
      description: Update specific fields of a trainee by their ID.
      parameters:
        - name: traineeId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the trainee to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TraineePatch'
            examples:
              updateName:
                summary: Update trainee name
                value:
                  name: "John Smith"
              updateEmail:
                summary: Update trainee email
                value:
                  email: "john.smith@example.com"
      responses:
        '200':
          description: Trainee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainee'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: [ ]
      tags:
        - Trainees
      summary: Delete a trainee
      description: Delete a specific trainee by their ID.
      parameters:
        - name: traineeId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the trainee to delete.
      responses:
        '204':
          description: Trainee deleted successfully
        '404':
          description: Trainee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions:
    post:
      tags:
        - Authentication
      summary: Create session (login)
      description: Authenticate trainee and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
            examples:
              login:
                summary: Login credentials
                value:
                  email: "john.doe@example.com"
                  password: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Authentication token
                  trainee:
                    $ref: '#/components/schemas/Trainee'
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                trainee:
                  id: "1"
                  name: "John Doe"
                  email: "john.doe@example.com"
                  timezone: "Europe/Tallinn"
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email and password are required"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid email or password"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Internal server error"

    delete:
      security:
        - BearerAuth: [ ]
      tags:
        - Authentication
      summary: Destroy session (logout)
      description: End trainee session
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Successfully logged out"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Unauthorized"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Internal server error"

    get:
      security:
        - BearerAuth: [ ]
      tags:
        - Authentication
      summary: Check session
      description: Check if trainee is authenticated
      responses:
        '200':
          description: Session status
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  trainee:
                    $ref: '#/components/schemas/Trainee'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /workouts:
    post:
      security:
        - BearerAuth: [ ]
      tags:
        - Workouts
      summary: Create a new workout type
      description: Create a new workout type with a name, duration, and optional details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutType'
            examples:
              workout:
                summary: Example workout
                value:
                  name: "High-Intensity Interval Training"
                  duration: 45
                  description: "Alternating periods of short intense exercise with less intense recovery periods"
                  color: "#FF5733"
      responses:
        '201':
          description: Workout type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutType'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid input"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      security:
        - BearerAuth: [ ]
      tags:
        - Workouts
      summary: List all workouts
      description: Retrieve a list of all workout types.
      responses:
        '200':
          description: List of workouts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workout'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /workouts/{workoutId}:
    get:
      security:
        - BearerAuth: [ ]
      tags:
        - Workouts
      summary: Get workout details
      description: Retrieve details of a specific workout by its ID.
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the workout to retrieve.
      responses:
        '200':
          description: Workout found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      security:
        - BearerAuth: [ ]
      tags:
        - Workouts
      summary: Partially update a workout
      description: Update specific fields of a workout by its ID.
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the workout to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutPatch'
            examples:
              updateName:
                summary: Update workout name
                value:
                  name: "Advanced HIIT"
              updateDuration:
                summary: Update workout duration
                value:
                  duration: 60
                  description: "Extended high-intensity workout"
      responses:
        '200':
          description: Workout updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutType'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: [ ]
      tags:
        - Workouts
      summary: Delete a workout
      description: Delete a specific workout by its ID.
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the workout to delete.
      responses:
        '204':
          description: Workout deleted successfully
        '404':
          description: Workout not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
      required:
        - page
        - pageSize
        - total
    Trainee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        timezone:
          type: string
        token:
          type: string  # Token is optional
      required:
        - id
        - name
        - email
        - password
    TraineePatch:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        timezone:
          type: string
      minProperties: 1  # At least one field must be provided
    WorkoutType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      required:
        - name
        - duration
    Workout:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - duration
    WorkoutPatch:
      type: object
      properties:
        name:
          type: string
        duration:
          type: integer
        description:
          type: string
        color:
          type: string
      minProperties: 1  # At least one field must be provided
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad Request"
      required:
        - message